STAR-11121 Add support for restore jobs
This is a basic implementation of restore job.
It supports only selected set of jobs parameters,
for example, it does take into account priority,
different manifest formats. It does not generate report
file at the end of the job. It does not check ETag
of manifest. Tests are also not implemented.
Supported opertions:
- creating a job
- getting description of a job

It is possible to change delay time for restoration
using the following endpoint from the internal API:

POST /moto-api/s3control/configure
{"restoration_delay_in_seconds": 123}